修正首页金额小数点问题：Rabbit中生成Redis时

See merge request rapserver/smarthelper!606

+++ common-lib/src/main/java/com/rpa/common/mapper/OrderMapper.java
@@ -22,7 +22,7 @@ public interface OrderMapper extends BaseMapper<OrderPO, OrderPOKey> {
 
     int queryPayCount();
 
-    float queryDayRevenue();
+    Long queryDayRevenue();
 
-    float queryMonthRevenue();
+    Long queryMonthRevenue();
 }
